+++ example/lib/example_kits/1_multi_state/10_multi_view_state_builder.dart
@@ -1,5 +1,5 @@
 import 'package:example/example_kits/providers/1_view_state_notifier.dart';
-import 'package:example/example_kits/providers/2_provider_kit.dart';
+import 'package:example/example_kits/providers/2_async_view_state_notifier.dart';
 import 'package:example/scaffold_with_button.dart';
 import 'package:flutter/material.dart';
 import 'package:provider_kit/provider_kit.dart';