+++ example/lib/example_kits/providers/2_async_view_state_notifier.dart
@@ -1,13 +1,13 @@
 import 'dart:async';
+
 import 'package:example/repository/repository.dart';
 import 'package:provider_kit/provider_kit.dart';
 
-class ItemsProvider extends ProviderKit<List<Item>> {
+class ItemsProvider extends AsyncViewStateNotifier<List<Item>> {
   final Repository _repo = Repository();
 
   @override
   FutureOr<List<Item>> fetchData() => _repo.getItems(10);
-
 }
 
 typedef ItemsViewState = ViewState<List<Item>>;